fix:Change duration data type to NUMERIC
Earlier there was no match when trying to match song duration with log_data because of float(5). So change the data type to NUMERIC for raw matching.

+++ sql_queries.py
@@ -38,7 +38,7 @@ CREATE TABLE IF NOT EXISTS songs (
     title varchar NOT NULL,
     artist_id varchar,
     year int NOT NULL,
-    duration float(5) NOT NULL
+    duration numeric NOT NULL
 );
 """)
